+++ core/src/main/kotlin/io/github/chrislo27/rhre3/entity/model/special/ShakeEntity.kt
@@ -18,7 +18,7 @@ class ShakeEntity(remix: Remix, datamodel: ShakeScreen)
     companion object {
 
         fun getShakeIntensity(semitone: Int): Float {
-            return Semitones.getALPitch(semitone)
+            return Semitones.getALPitch(semitone.toFloat())
         }
 
     }
